CLeanup toMapTree typing
