+++ packages/dds/tree/src/test/simple-tree/mapTree.spec.ts
@@ -36,8 +36,6 @@ import {
 	stringSchema,
 	type TreeNodeSchema,
 } from "../../simple-tree/index.js";
-// eslint-disable-next-line import/no-internal-modules
-import type { InsertableContent } from "../../simple-tree/proxies.js";
 import {
 	type ContextualFieldProvider,
 	type ConstantFieldProvider,
@@ -52,6 +50,7 @@ import {
 	addDefaultsToMapTree,
 	getPossibleTypes,
 	mapTreeFromNodeData,
+	type InsertableContent,
 	// eslint-disable-next-line import/no-internal-modules
 } from "../../simple-tree/toMapTree.js";
 import { brand } from "../../util/index.js";
